day3/requirements.txt
gradio
chromadb
sentence-transformers
langchain-groq
pypdf
python-dotenv
